[New] Converted test cases for incidents from case components from enzyme to rtl
