+++ app/javascript/components/incidents-from-case/components/panel/component.spec.js
@@ -1,18 +1,11 @@
+import { mountedComponent, screen } from "test-utils";
 import { fromJS } from "immutable";
-import { Accordion, AccordionDetails, AccordionSummary } from "@material-ui/core";
-
-import LookupValue from "../../../record-form/form/subforms/subform-header-lookup";
-import { setupMountedComponent } from "../../../../test";
-import IncidentSummary from "../summary";
-import IncidentDetail from "../detail";
 import { RECORD_TYPES } from "../../../../config";
 import * as R from "../../../record-form/records";
 import { mapEntriesToRecord } from "../../../../libs";
-
 import IncidentPanel from "./component";
 
 describe("<IncidentPanel /> - Component", () => {
-  let component;
   const props = {
     incident: fromJS({
       created_by: "primero_gbv",
@@ -27,8 +20,8 @@ describe("<IncidentPanel /> - Component", () => {
     incidentCaseId: "case-id-1",
     css: {},
     mode: { isShow: false, isEdit: true },
-    setFieldValue: () => {},
-    handleSubmit: () => {},
+    setFieldValue: () => { },
+    handleSubmit: () => { },
     recordType: RECORD_TYPES.cases
   };
 
@@ -104,46 +97,27 @@ describe("<IncidentPanel /> - Component", () => {
   });
 
   beforeEach(() => {
-    ({ component } = setupMountedComponent(IncidentPanel, props, initialState));
+    mountedComponent(<IncidentPanel {...props} />, initialState);
   });
 
   it("render IncidentPanel component", () => {
-    expect(component.find(IncidentPanel)).to.have.length(1);
+    expect(screen.getByRole("panel")).toBeInTheDocument();
   });
 
-  it("render a Accordions", () => {
-    expect(component.find(Accordion)).to.have.lengthOf(1);
-    expect(component.find(AccordionSummary)).to.have.lengthOf(1);
-    expect(component.find(AccordionDetails)).to.have.lengthOf(1);
+  it("render IncidentSummary component", () => {
+    expect(screen.getByText("incidents.type_violence")).toBeInTheDocument();
   });
 
-  it("render a IncidentSummary", () => {
-    expect(component.find(IncidentSummary)).to.have.lengthOf(1);
+  it("render IncidentDetail component", () => {
+    expect(screen.getByText("incidents.date_of_incident")).toBeInTheDocument();
   });
 
-  it("render a IncidentDetail", () => {
-    expect(component.find(IncidentDetail)).to.have.lengthOf(1);
-  });
-
-  describe("with violence-type-lookup", () => {
-    it("should use the lookup defined in the option_strings_source", () => {
-      expect(component.find(IncidentDetail).find(LookupValue).props().optionsStringSource).to.equal(
-        "lookup-gbv-sexual-violence-type"
-      );
-    });
-
-    it("renders the translated value", () => {
-      expect(component.find(IncidentDetail).find(LookupValue).text()).to.equal("Test1");
-    });
-  });
-
-  it("renders component with valid props", () => {
-    const incidentsProps = { ...component.find(IncidentPanel).props() };
-
-    ["incident", "incidentCaseId", "css", "mode", "setFieldValue", "handleSubmit", "recordType"].forEach(property => {
-      expect(incidentsProps).to.have.property(property);
-      delete incidentsProps[property];
-    });
-    expect(incidentsProps).to.be.empty;
+  it("with violence-type-lookup-renders the translated value", () => {
+    expect(screen.getAllByText("Test1")).toBeTruthy();
   });
 });
+
+
+
+
+
